+++ src/api/qualicharge/api/v1/__init__.py
@@ -1,18 +1,19 @@
 """QualiCharge API v1."""
 
 import logging
-from typing import Annotated, Union
+from typing import Union
 
 from fastapi import FastAPI, Request, Security, status
 from fastapi.responses import JSONResponse
 
-from qualicharge.auth.models import IDToken, User
 from qualicharge.auth.oidc import get_token
 from qualicharge.exceptions import OIDCAuthenticationError, OIDCProviderException
 
+from .routers import auth, statique
+
 logger = logging.getLogger(__name__)
 
-app = FastAPI(title="QualiCharge API (v1)")
+app = FastAPI(title="QualiCharge API (v1)", dependencies=[Security(get_token)])
 
 
 @app.exception_handler(OIDCAuthenticationError)
@@ -27,8 +28,5 @@ async def authentication_exception_handler(
     )
 
 
-@app.get("/whoami")
-async def me(token: Annotated[IDToken, Security(get_token)]) -> User:
-    """A test endpoint to validate user authentication."""
-    logger.debug(f"{token=}")
-    return User(email=token.email)
+app.include_router(auth.router)
+app.include_router(statique.router)